+++ src/lib/reporters/phpunit-10/src/Status.php
@@ -21,13 +21,21 @@ enum Status: string
     {
         return match (get_class($event)) {
             Test\ConsideredRisky::class => self::EMPTY,
+            Test\DeprecationTriggered::class => self::WARNING,
             Test\Errored::class => self::FAILED,
-            Test\Failed::class => self::FAILED,
+            Test\ErrorTriggered::class => self::WARNING,
             Test\Failed::class => self::FAILED,
             Test\MarkedIncomplete::class => self::INCOMPLETE,
+            Test\NoticeTriggered::class => self::WARNING,
             Test\Passed::class => self::PASSED,
-            Test\Skipped::class => self::SKIPPED,
+            Test\PhpDeprecationTriggered::class => self::WARNING,
+            Test\PhpNoticeTriggered::class => self::WARNING,
             Test\PhpunitDeprecationTriggered::class => self::WARNING,
+            Test\PhpunitErrorTriggered::class => self::WARNING,
+            Test\PhpunitWarningTriggered::class => self::WARNING,
+            Test\PhpWarningTriggered::class => self::WARNING,
+            Test\Skipped::class => self::SKIPPED,
+            Test\WarningTriggered::class => self::WARNING,
         };
     }
 }