+++ src/lib/reporters/phpunit-10/src/Subscriber/TestPhpDeprecationTriggeredSubscriber.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LodeApp\PHPUnit\Subscriber;
+
+use PHPUnit\Event\Test\PhpDeprecationTriggered;
+use PHPUnit\Event\Test\PhpDeprecationTriggeredSubscriber;
+
+final class TestPhpDeprecationTriggeredSubscriber extends Subscriber implements PhpDeprecationTriggeredSubscriber
+{
+    public function notify(PhpDeprecationTriggered $event): void
+    {
+        $this->reporter()->handleEvent($event);
+    }
+}